requirements.txt
amqp==5.3.1
asgiref==3.9.2
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.6
django-auditlog==3.3.0
django-cors-headers==4.9.0
django-debug-toolbar==6.0.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
geopandas==1.1.1
idna==3.10
inflection==0.5.1
kombu==5.5.4
Markdown==3.9
networkx==3.5
numpy==2.3.3
osmnx==2.0.6
packaging==25.0
pandas==2.3.3
pillow==11.3.0
prompt_toolkit==3.0.52
PyJWT==2.10.1
pyogrio==0.11.1
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
shapely==2.1.2
six==1.17.0
sqlparse==0.5.3
stripe==12.5.1
timedelta==2020.12.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
